feat(go-keychain): cross-platform keychain with build tags
Split keychain into platform-specific files using Go build tags:
- macOS: security CLI (unchanged behavior)
- Linux: secret-tool CLI (libsecret)

No CGo required, pure Go on both platforms.
